OCTO-25 Subscription

+++ bin/cleanup.sh
@@ -10,7 +10,10 @@ for f in $(find "$ROOT_DIR" -name "*.pyc" -type f); do
 done
 # delete sass cache
 echo "[INFO] Removing sass-cache files in $ROOT_DIR"
-rm -r "$ROOT_DIR/.sass-cache"
+if [ -d "$ROOT_DIR/.sass-cache" ]; then
+    echo "- Removing directory"
+    rm -r "$ROOT_DIR/.sass-cache"
+fi
 # delete service logs
 echo "[INFO] Removing 'octohaven-service.log' files"
 for f in $(find "$ROOT_DIR" -name "octohaven-service.log*" -type f); do
